contract/requirements.txt
appdirs==1.4.3
asgiref==3.2.3
astroid==2.3.3
attrs==19.3.0
Automat==20.2.0
backcall==0.1.0
baidu-aip==2.2.18.0
beautifulsoup4==4.9.0
bpmappers==1.1
bs4==0.0.1
certifi==2019.9.11
cffi==1.14.0
chardet==3.0.4
ci-info==0.2.0
click==7.1.2
colorama==0.4.1
comtypes==1.1.7
configobj==5.0.6
configparser==5.0.1
constantly==15.1.0
cryptography==2.9.2
cssselect==1.1.0
decorator==4.4.1
distlib==0.3.0
Django==3.0.1
django-cors-headers==3.2.1
django-debug-toolbar==2.2
etelemetry==0.2.2
filelock==3.0.12
fitz==0.0.1.dev2
Flask==1.1.2
future==0.18.2
gevent==20.5.0
greenlet==0.4.15
httplib2==0.18.1
hyperlink==19.0.0
idna==2.8
imageio==2.8.0
imageio-ffmpeg==0.4.1
incremental==17.5.0
ipython==7.9.0
ipython-genutils==0.2.0
isodate==0.6.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.11.2
lazy-object-proxy==1.4.3
lxml==4.5.0
MarkupSafe==1.1.1
mccabe==0.6.1
mongoengine==0.20.0
moviepy==1.0.2
networkx==2.5
nibabel==3.1.1
nipype==1.5.1
numpy==1.17.4
opencv-python==4.1.2.30
packaging==20.4
pandas==1.1.3
parsel==1.5.2
parso==0.5.1
pickleshare==0.7.5
Pillow==7.1.1
proglog==0.1.9
prompt-toolkit==2.0.10
Protego==0.1.16
prov==1.5.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
pydot==1.4.1
pydotplus==2.0.2
Pygments==2.4.2
PyHamcrest==2.0.2
pylint==2.4.4
pymongo==3.10.1
PyMuPDF==1.18.1
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2019.3
pywin32 @ file:///C:/Users/Les%20Grand%20Bleu/Downloads/pywin32-228-cp38-cp38-win32.whl
pyxnat==1.3
qrcode==6.1
queuelib==1.5.0
rdflib==5.0.0
requests==2.22.0
scipy==1.5.3
Scrapy==2.1.0
selenium==3.141.0
service-identity==18.1.0
simplejson==3.17.2
six==1.13.0
soupsieve==2.0
sqlparse==0.3.0
tencentcloud-sdk-python==3.0.112
tqdm==4.46.0
traitlets==4.3.3
traits==6.1.1
Twisted==20.3.0
urllib3==1.25.7
virtualenv==20.0.18
w3lib==1.21.0
wcwidth==0.1.7
wechatsogou==4.5.4
Werkzeug==0.16.1
wrapt==1.11.2
xlrd==1.2.0
XlsxWriter==1.2.8
xlwt==1.3.0
zhon==1.1.5
zope.interface==5.1.0
